+++ rotkehlchen/chain/evm/decoding/merkl/constants.py
@@ -1,0 +1,6 @@
+from typing import Final
+
+from rotkehlchen.chain.evm.types import string_to_evm_address
+
+CPT_MERKL: Final = 'merkl'
+MERKL_DISTRIBUTOR_ADDRESS: Final = string_to_evm_address('0x3Ef3D8bA38EBe18DB133cEc108f4D14CE00Dd9Ae')  # noqa: E501